chore: move run now handler to scrapers package
